Made the flashlight an EntityComponent
